Properly integrate CRLs into the TLS flow
When establishing a connection the CRL distribution points are
extracted from the upstream cert. They are passed to the dummy
certificate creation function and their original path is replaced
with the CA's serial number. This makes it possible to detect the
modified URL without having to store any additional per-URL state.

The certificate store has been extended to generate an empty CRL
at initialization. The tlsconfig addon has been extended to detect
a modified CRL URL and serve the empty CRL from the certificate
store.